Remove unnecessary addition of derived models to effective type candidates
It turns out that code I put in to replace the isDerivedFrom checks in the
prior implementation of getEffectiveType isn't needed. This is great because
I was going back to it because I realized it should have been recursive, and
to put a comment where code would need to change with property overrides but
then I ended up deleting the code. :)

Also added an additional test with another level of inheritance for good measure.